requirements.txt
openai
anthropic
google-genai
requests
pydantic>=2.0
pandas
tqdm
python-dotenv
